Re-implementing as.integer(NA).  Matrix/Algebra pointers (MxIndex, to the back-end) are now Integer slots.  That means NAs placed in those slots should be as.integer(NA).
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@513 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxMLObjective.c
@@ -196,7 +196,7 @@ void omxInitMLObjective(omxObjective* oo, SEXP rObj, SEXP dataList) {
 	omxMLObjective *newObj = (omxMLObjective*) R_alloc(1, sizeof(omxMLObjective));
 	
 	PROTECT(nextMatrix = GET_SLOT(rObj, install("means")));
-	if(!R_FINITE(REAL(nextMatrix)[0])) {
+	if(!R_FINITE(INTEGER(nextMatrix)[0])) {
 		if(OMX_DEBUG) {
 			Rprintf("ML: No Expected Means.\n");
 		}
